@smooth fixes, add library example
